Incorrect attempt at day 16 part 2
I mistakenly thought that you could just combine the permutations that
happened into a set of position permutations.  You can't because the
"partner" move is based on the name of the program, instead of the
position of the program.

So this approach doesn't work, but I've recorded it for interest.