A successful solution to day 16 part 2
This works by splitting the permutations of the names of the programs
and the permutations of the positions of the programs up.  It then
applied the dance 10 times, to work out what the permutation after 10
is.  It then applies this combined permutation 10 times, to work out the
permutation after 100 times.  It does this further times, to work out
the permutation after each power of ten up to a billion.

This does work, but is massively overkill, because it turns out the
dance has a pretty low period.