Day 20
This is the solution to part 2 - needs removing the collision code to
show the answer for part 1.

I made two errors on the way to getting here.

First was that my initial "dist" function just did the sum of the
distances without taking the "abs" of each direction.  Had to fix and
resubmit, but still got rank 136 for part 1.

Second was that in my haste I submitted the particle number that was
left closest to origin for part 2, rather than the number of particles
left, so had to wait a minute for that one, too. Was still fast enough
to get rank 59, which I'm pretty pleased with!