First half easy - second half caught me out a few times.
Firstly, forgot that the problem might be fixed by deleting the item either
side of the place where I found a problem.

Secondly, tried to remove the item on either side and retest, but forgot that
the first item determining the problem might be the one before either of those.
Rather than fixing neatly, just brute-forced trying deleting each item in the
row and retesting.