Figured out roughly what the program was doing.
It's shifting and bit flipping, but each output is entirely determined
by 10 bits of a; the first output is determined by the low 10 bits
(0-9), the second by bits (3-12), etc.

Search for the low 10 bits that emit the right output for the first
number.  then add all 3 bit combinations possible to that, and see which
produce the correct first two numbers. etc